+++ packages/delete/.eslintignore
