+++ packages/delete/.eslintrc.js
