+++ packages/delete/.npmignore
