+++ packages/delete/.prettierrc.js
