+++ packages/delete/jest.config.ts
