+++ packages/delete/package.json
