+++ packages/delete/tsconfig.json
