--- upload/models.ts
@@ -1,10 +1,0 @@
-/**
- * Represents package metadata
- */
-
-export interface Package {
-    name: string;
-    version: string;
-    repository: string;
-    binaryData: Buffer;
-}
